included modularity of the code
