+++ src/analysis/__init__.py
@@ -1,0 +1,22 @@
+"""
+Analysis module for data analysis operations.
+
+This module contains specialized analysis classes for different signal types,
+as well as the main Analysis class that coordinates analysis across multiple datasets.
+"""
+
+from .base_analysis import BaseAnalysis
+from .purity_analysis import PurityAnalysis
+from .temperature_analysis import TemperatureAnalysis
+from .h2o_analysis import H2OConcentrationAnalysis
+from .liquid_level_analysis import LiquidLevelAnalysis
+from .main_analysis import Analysis
+
+__all__ = [
+    'BaseAnalysis',
+    'PurityAnalysis',
+    'TemperatureAnalysis',
+    'H2OConcentrationAnalysis',
+    'LiquidLevelAnalysis',
+    'Analysis'
+]
