+++ src/core/__init__.py
@@ -1,0 +1,8 @@
+"""
+Core module containing fundamental data structures and base classes.
+"""
+
+from .standard_format import StandardDataFormat
+from .base_classes import BaseDataProcessor
+
+__all__ = ['StandardDataFormat', 'BaseDataProcessor']
