+++ src/utils/__init__.py
@@ -1,0 +1,12 @@
+"""
+Utility module for common functionality.
+
+This module contains utility classes and functions used throughout
+the framework, such as data validation and other helper functions.
+"""
+
+from .data_validator import DataValidator
+
+__all__ = [
+    'DataValidator'
+]
